More permissively translate HTML
The current logic that decides wether HTML should be translated
will often not translate text in an EPUB that the user would
want translated. This commite will translate any html or xhtml
files in the EPUB document.

Example of a filename that isn't translated with the current code
but is with this commit:
OEBPS/sections/section0003.xhtml